// src/store/emailTemplate/emailTemplate.enum.ts
enum EmailTemplateTypeEnum {
	GET_EMAIL_TEMPLATE_SUCCESS = "GET_EMAIL_TEMPLATE_SUCCESS",
	GET_EMAIL_TEMPLATE_PENDING = "GET_EMAIL_TEMPLATE_PENDING",
	GET_EMAIL_TEMPLATE_FAILED = "GET_EMAIL_TEMPLATE_FAILED",

	GET_SINGLE_EMAIL_TEMPLATE_SUCCESS = "GET_SINGLE_EMAIL_TEMPLATE_SUCCESS",
	GET_SINGLE_EMAIL_TEMPLATE_PENDING = "GET_SINGLE_EMAIL_TEMPLATE_PENDING",
	GET_SINGLE_EMAIL_TEMPLATE_FAILED = "GET_SINGLE_EMAIL_TEMPLATE_FAILED",

	UPDATE_EMAIL_TEMPLATE_SUCCESS = "UPDATE_EMAIL_TEMPLATE_SUCCESS",
	UPDATE_EMAIL_TEMPLATE_PENDING = "UPDATE_EMAIL_TEMPLATE_PENDING",
	UPDATE_EMAIL_TEMPLATE_FAILED = "UPDATE_EMAIL_TEMPLATE_FAILED",
}
export default EmailTemplateTypeEnum;